+++ grammars/runtime/builtin_scanners_test.go
@@ -1,0 +1,5 @@
+//go:build !grammar_subset
+
+package grammarruntime
+
+func init() { RegisterBuiltinScanners() }