+++ grammars/z_subset_token_source_factory_authzed.go
@@ -1,7 +1,0 @@
-//go:build grammar_subset && grammar_subset_authzed
-
-package grammars
-
-func init() {
-	registerTokenSourceFactory("authzed", NewAuthzedTokenSourceOrEOF)
-}