+++ grammars/z_subset_token_source_factory_c_family.go
@@ -1,8 +1,0 @@
-//go:build grammar_subset && (grammar_subset_c || grammar_subset_cpp)
-
-package grammars
-
-func init() {
-	registerTokenSourceFactory("c", NewCTokenSourceOrEOF)
-	registerTokenSourceFactory("cpp", NewCTokenSourceOrEOF)
-}